feat: saving mock csv conformational b prediction data from fastapi
