--- apps/nextjs/src/app/(app)/account/[[...account]]/user-settings.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import type { User } from "@supabase/supabase-js";
-
-import { getNameFromUser } from "~/lib/utils";
-
-export default function UserSettings({ user }: { user: User }) {
-  return (
-    <>
-      <p>{getNameFromUser(user)}</p>
-      <p>{user.email ?? ""}</p>
-    </>
-  );
-}
